Options -
Colour
Shape
Size
Layout
There should be three sections in the MAgento 2 admin (as there is in the M1 admin) to set these options for three different locations. These locations are

The mini-cart
The cart page
The checkout page
The options should use the same default options as the Magento 1 module.
Please ignore the Product Page, as this is not functionality supported by the M2 module.