float getPrice due to inconsistent return
getPrice() should return a float, or null, according to /vendor/magento/module-sales/Api/Data/OrderItemInterface.php but when setting a custom price from the admin, it is returned as a string, breaking the Level 2/3 Processing data builder class.

getBasePrice() has a similar issue if fixed price levels are set at the product level.

Both these issues should be fixed in Magento 2.3.5 but for now we should just use getPrice() and force the float.